Refactor towers tab into dedicated module
